Move vufs dto to common-utils module.
