+++ common-utils/src/main/java/com/stnetix/ariaddna/commonutils/dto/CertificateDTO.java
