+++ storage-service/pom.xml
@@ -55,11 +55,6 @@
             <artifactId>common-utils</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
-        <dependency>
-            <groupId>com.stnetix.ariaddna</groupId>
-            <artifactId>vufs</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
 
     </dependencies>
     <build>
