+++ storage-service/src/main/java/com/stnetix/ariaddna/persistence/services/IMetatableService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017 stnetix.com. All Rights Reserved.
+ * Copyright (c) 2018 stnetix.com. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License.  You may obtain a copy of
@@ -13,7 +13,7 @@
 
 package com.stnetix.ariaddna.persistence.services;
 
-import com.stnetix.ariaddna.vufs.dto.MetatableDTO;
+import com.stnetix.ariaddna.commonutils.dto.vufs.MetatableDTO;
 
 /**
  * 26.02.18.
